frontend/requirements.txt
jinja2
uplink
xrpl-py
